pyproject.toml
[project]
name = "xcountry-scanner"
version = "0.1.0"
description = "A Barcode scanner with a counter."
authors = [
    {name = "bohrium2b",email = "dzm2026@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django (>=5.2.1,<6.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
